backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
lightgbm==4.3.0
shap==0.45.0
joblib==1.4.2
openai==1.30.0
python-dotenv==1.0.1
ta==0.11.0
